+++ src/control.py
@@ -1,0 +1,56 @@
+################################################################################
+# Copyright (C) 2012-2013 Leap Motion, Inc. All rights reserved.               #
+# Leap Motion proprietary and confidential. Not for distribution.              #
+# Use subject to the terms of the Leap Motion SDK Agreement available at       #
+# https://developer.leapmotion.com/sdk_agreement, or another agreement         #
+# between Leap Motion and you, your company or other organization.             #
+################################################################################
+import os, sys, inspect
+src_dir = os.path.dirname(inspect.getfile(inspect.currentframe()))
+arch_dir = '../lib/x64' if sys.maxsize > 2**32 else '../lib/x86'
+sys.path.insert(0, os.path.abspath(os.path.join(src_dir, arch_dir)))
+
+import Leap, sys
+
+
+class SampleListener(Leap.Listener):
+
+    def on_frame(self, controller):
+        # Get the most recent frame and report some basic information
+        frame = controller.frame()
+        if len(frame.hands) == 2:
+            for hand in frame.hands:
+                if hand.is_left:
+                    leftHeight = hand.palm_position.y
+                    leftForward = hand.palm_position.z
+                else:
+                    rightHeight = hand.palm_position.y
+                    rightForward = hand.palm_position.z
+                try:
+                    print "Turn: " + str(leftHeight-rightHeight)
+                    print "Throttle: " + str((leftForward+rightForward)/2)
+                except:
+                    pass
+        else:
+            print "0"
+
+def main():
+    # Create a sample listener and controller
+    listener = SampleListener()
+    controller = Leap.Controller()
+
+    # Have the sample listener receive events from the controller
+    controller.add_listener(listener)
+
+    # Keep this process running until Enter is pressed
+    try:
+        sys.stdin.readline()
+    except KeyboardInterrupt:
+        pass
+    finally:
+        # Remove the sample listener when done
+        controller.remove_listener(listener)
+
+
+if __name__ == "__main__":
+    main()
